Update `ISettingRegistry.ISettings` interface to support querying of default values for root schema keys.
